// src/CommandBlock.Frontend/src/app/api/model/backupScheduleDto.ts
/**
 * CommandBlock.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BackupScheduleDto { 
    id: string;
    serverId: string;
    cronExpression: string;
    enabled: boolean;
    nextRunAt?: string | null;
    lastRunAt?: string | null;
    lastStatus?: string | null;
    lastError?: string | null;
    createdAt: string;
}